Updated documentation and reorganized repo
