+++ src/constants/README.md
@@ -1,0 +1,5 @@
+# 🗂 Constants
+Contains constant values that will be re-used throughout the platform.
+
+## ✏️ Notes
+Colors and the player rating constants are defined here so that they can be easily changed in one place down the line.
