+++ src/screens/welcome-screen.tsx
@@ -8,7 +8,7 @@ import { emailSchema } from 'src/utils/schemas';
 import { FormInput } from 'src/components/user-input';
 import { KeyboardBehaviorWrapper } from 'src/components/wrappers';
 import { anonymousSignIn, fetchSignInMethods } from 'src/firebase/api';
-import { useAppSelector, useAppDispatch } from 'src/hooks/useful-ducks';
+import { useAppDispatch, useAppSelector } from 'src/ducks/useful-hooks';
 import { guestSignIn } from 'src/ducks/user-slice';
 import { AuthStackParams } from 'src/navigation/auth-stack';
 import MaetSvg from 'assets/MaetSvg.svg';
