+++ src/utils/README.md
@@ -1,0 +1,12 @@
+# 🗂 Utils
+Various utility files used throughout the repo.
+
+## ✏️ Notes
+We have defined many of the ways we handle dates, ratings, our verification, and yup schemas in this directory.
+
+### 🟥 Concerns or Unknown Functionality
+1. What sort of utilities should we have the database handle, versus the client, especially as it relates to ratings and date.
+
+## ↔️ Closely Related Directories
+
+## 🔗 Resources
